requirements.txt
schedule
slackclient
coverage>=4.5
coveralls
